Render Retirements Trends tabs with SSR
