+++ carbonmark-data/components/pages/RetirementTrendsPage/styles.module.scss
@@ -1,0 +1,24 @@
+.text {
+  display: grid;
+  grid-template-rows: 1fr 1fr;
+
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: 6rem;
+
+  /**
+  @todo import the breakpoint for desktopLarge here
+  */
+  @media (min-width: 75em) {
+    grid-template-columns: 1fr 1fr;
+    width: 80rem;
+  }
+}
+
+.insights {
+  h2 {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+}
